FEATURE: Recently used emotes
New enhancement feature: Option to show recently used emotes in a custom
tab in the YouTube emote picker window.
Relies heavily on YouTubes infrastructure for emotes and the picker
window, and for now there is no limit for the number of emotes that
could appear there. Marked as BETA-feature for now, to be used with
awareness for possible bugs or performance issues.